tensorrt-integrate-1.5-unet/unet-pytorch-change/requirements.txt
scipy
numpy
matplotlib
opencv_python
torch
torchvision
tqdm
Pillow
h5py